ep transaction complete

+++ src/models/epTransaction.go
@@ -4,8 +4,7 @@ import "gorm.io/gorm"
 
 type EpTransaction struct {
 	gorm.Model
-	DistribId string `json:"distrib_id"`
-	Reference string `json:"reference"`
-	Value     int    `json:"value"`
+	DistribId string  `json:"distrib_id"`
+	Reference string  `json:"reference"`
+	Value     float64 `json:"value"`
 }
-
